refactor(police): extract useCanvasArrange — keep App.tsx as a thin shell
\`.claude/rules/solidjs.md\` is explicit: "Keep App.tsx as a thin
layout shell. State per domain: extract shared state into useXxx.ts
modules." The arrange feature accumulated five helpers in App.tsx —
\`applyTileGeometry\`, \`getBucketFor\`, \`repoIslandTileFor\`,
\`resolvePlacementBucket\`, \`handleCanvasAutoArrange\` — plus an
inline \`placeNew\` closure on the \`<TerminalCanvas>\` props. That's
the kind of overload the rule warns against.

Move the lot to \`canvas/useCanvasArrange.ts\`. The hook takes
\`{ store, crud, viewport, isMobile }\` and returns three callables
(\`applyTileGeometry\`, \`placeNew\`, \`handleCanvasAutoArrange\`).
App.tsx shrinks back to its layout-shell shape — the canvas's three
arrange-touching props become \`arrange.placeNew\`,
\`arrange.applyTileGeometry\`, \`arrange.handleCanvasAutoArrange\`.

No behavior change. 23/23 canvas e2e + 11/11 unit tests pass.
